Moved the library name methods into WebPFactory
I've moved the stuff for the library name and additional search paths into WebPFactory since I didn't realize that JNRFFI was a package exclusive class.
